requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arize-phoenix-client==1.4.0
arize-phoenix-otel==0.9.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
openai==1.77.0
openinference-instrumentation==0.1.28
openinference-instrumentation-openai==0.1.27
openinference-semantic-conventions==0.1.17
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==25.0
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.21.0
